move/rename files for consistent format, remove comment, change url
