+++ src/commands/index.ts
@@ -9,7 +9,7 @@ export * from "./run";
 export * from "./stop";
 export * from "./runmedic";
 export * from "./fw-update";
-export * from "./prosDocs";
+export * from "../views/docs-webscrape";
 export * from "./command_tools";
 export * from "./base-command";
 export * from "./set-team-number";
