+++ src/views/docs-webscrape.ts
@@ -13,7 +13,7 @@ var prosJson = {
     "members": [
       {
         "name": "Motor",
-        "url": "https://purduesigbots.github.io/pros-doxygen-docs/group__cpp-motors.html",
+        "url": "https://purduesigbots.github.io/pros3-doxygen-docs/group__cpp-motors.html",
         "functions": [{"name" : "move", "url" : ""}]
       }
       ],
@@ -22,7 +22,7 @@ var prosJson = {
       "members": [
         {
           "name": "Motor",
-          "url": "https://purduesigbots.github.io/pros-doxygen-docs/group__c-motors.html",
+          "url": "https://purduesigbots.github.io/pros3-doxygen-docs/group__c-motors.html",
           "functions": [{"name" : "move", "url" : ""}]
         }
       ]
@@ -70,8 +70,8 @@ export function parseJSON(keyword: string){
  async function loadJson(){
   const axiosInstance = axios.create(); // Create a new Axios Instance
  
-  const url = 'https://purduesigbots.github.io/pros-doxygen-docs/api.html'; // PROS Doxygen url
-  const prosBaseUrl = "https://purduesigbots.github.io/pros-doxygen-docs/";
+  const url = 'https://purduesigbots.github.io/pros3-doxygen-docs/api.html'; // PROS Doxygen url
+  const prosBaseUrl = "https://purduesigbots.github.io/pros3-doxygen-docs/";
   var variable: any;
   var finalLink = "";
   const linkRegex = /<a href="(.{0,30})" class="m-doc">(.{0,35})<\/a>/g;
